requirements.txt
pypdf
cohere
tiktoken
sentence-transformers
pinecone-client
langchain-community
langchain-pinecone
langchain_openai
langchain
openai
streamlit
